Extract sync cycle runner
