+++ src/lib.rs
@@ -49,6 +49,7 @@ mod setup;
 mod shutdown;
 mod state;
 mod string_interner;
+mod sync_cycle;
 mod sync_loop;
 mod systemd;
 mod types;
